refactor: finalize component extract refactoring
