+++ public/global.css
@@ -1,3 +1,9 @@
+:root {
+  --font-regular: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
+    Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif;
+  --font-monospace: "Courier New", Courier, monospace;
+}
+
 /* See https://www.30secondsofcode.org/css/s/box-sizing-reset */
 *,
 *::before,
@@ -13,10 +19,11 @@ body {
 }
 
 body {
-  color: #333;
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-    Oxygen-Sans, Ubuntu, Cantarell, "Helvetica Neue", sans-serif;
+  background-image: url("https://upload.wikimedia.org/wikipedia/commons/f/f3/43228_1_miscellaneous_digital_art_simple_simple_patterns.jpg");
+  background-size: cover;
+  font-family: var(--font-monospace);
+  color: white;
 }
 
 a {
@@ -40,15 +47,16 @@ button,
 input,
 select,
 textarea {
+  background-color: transparent;
+  color: inherit;
   font-size: 1rem;
 }
 
 input,
 select,
 textarea {
-  font-family: inherit;
-  padding: 0.5rem;
-  margin: 0 0 0.5rem 0;
+  padding: 0.75rem;
+  margin: 0;
   border: 1px solid #ccc;
   border-radius: 2px;
 }
@@ -58,8 +66,6 @@ input:disabled {
 }
 
 button {
-  color: white;
-  background-color: transparent;
   border: none;
 }
