+++ src/Footer.svelte
@@ -1,52 +1,37 @@
 <script>
-  import { createEventDispatcher } from "svelte";
-
+  export let alertsOn;
   export let count;
   export let mode;
-  export let alertsOn;
+  export let tasks;
 
-  const dispatchEvent = createEventDispatcher();
-
-  // function to filter active tasks
-  function filterActiveTasks() {
-    mode = "active";
-  }
-
-  // function to filter all tasks
-  function filterAllTasks() {
-    mode = "all";
-  }
-
-  // function to filter active tasks
-  function filterCompletedTasks() {
-    mode = "completed";
-  }
-
-  function handleClearClick() {
-    dispatchEvent("clear");
+  // function to clear all tasks
+  function clearAllTasks() {
+    if (tasks.length !== 0) {
+      tasks = [];
+      if (alertsOn) {
+        alert("All done! Time for a drink! 🍸");
+      }
+    }
   }
 </script>
 
 <footer>
   <span>
-    {#if count === 1}
-      {count} task remaining
-    {:else}
-      {count} tasks remaining
-    {/if}
+    {count}
+    {count === 1 ? "task" : "tasks"} remaining
   </span>
   <span>|</span>
   <span>
     Showing {mode} tasks
   </span>
   <span>|</span>
-  <button on:click={filterAllTasks}>All</button>
+  <button on:click={() => (mode = "all")}>All</button>
   <span>|</span>
-  <button on:click={filterActiveTasks}>Active</button>
+  <button on:click={() => (mode = "active")}>Active</button>
   <span>|</span>
-  <button on:click={filterCompletedTasks}>Completed</button>
+  <button on:click={() => (mode = "completed")}>Completed</button>
   <span>|</span>
-  <button on:click={handleClearClick}>Clear All Tasks</button>
+  <button on:click={clearAllTasks}>Clear All Tasks</button>
   <span>|</span>
   <button on:click={() => (alertsOn = !alertsOn)}>Toggle Alerts</button>
 </footer>
@@ -54,10 +39,8 @@
 <style>
   footer {
     position: fixed;
-    left: 0;
-    bottom: 0;
+    inset: auto 0 0 0;
     width: 100%;
-    text-align: center;
     display: flex;
     align-items: center;
     justify-content: center;
@@ -67,5 +50,6 @@
   button,
   span {
     padding: 0.5rem;
+    font-family: var(--font-regular);
   }
 </style>
